assert listagem, or vgr ger

+++ AscFrontEnd/DTOs/Compra/CotacaoArtigoDTO.cs
@@ -6,6 +6,6 @@
         public float preco { get; set; }
         public float iva { get; set; }
         public float desconto { get; set; }
-        public int qtd { get; set; }
+        public float qtd { get; set; }
     }
 }
